Initial commit.
Example Processing program that extracts the mean and standard
deviation of incoming sensor data (by some time window), allows
association of class labels (0 to 9) with data from the latest time
window, recognizes the class of the closest neighbor of the latest time
window, and generates a data file for an Arduino program that does
similar recognition.

This is a proof-of-concept example, with canned features and
classification algorithm.