Add overview of test data.
To allow zooming into a particular window. This requires caching
predictions on the test data, so we don’t have to re-run prediction
every time we change the window of test data shown.

I also updated Plotter so that it recalculates minY and maxY when you
call setData() (by successively calling push_back instead of just
assigning the data matrix).