Show snapshots of high-dimensional training samples.
If the number of dimensions of our data is greater than the
kTooManyFeaturesThreshold threshold, show a snapshot for each training
sample, along with an overview of the whole sample. Note that we still
need to allow for scrubbing through the sample (i.e. showing the
snapshot at each moment in time).