Use separate range for each class distance plot.
For ANBC classifiers, the ranges of the plots can be dramatically
different depending on the standard deviation of the class’s training
data. As a result, using the same range for every class makes some
plots impossible to read (collapsed to the x-axis). Using a different
range for each class makes it harder to compare the plots to each
other, but it’s more important to be able to compare each class’s
distance to the threshold.