fix(ofApp): explicitly enable/disable buttons
It turns out that if a widget is not shown, it's listener is still
functioning (i.e. mouse events are still captured). This happens to both
ofxDatGui as well as our own plotter. For now, I am explicitly setting
ofxDatGui components enable/disable. We are really not using
openFrameworks in an easy way...

References #347