fix(rename/relabel): mouse release makes it stop
In MouseRelease handler, if we are in the middle of renaming or
relabelling, we stop it and return to AppState::kTraining state. Note
because of this change, the variable name new_state is confusing,
potential_new_state might be better, but still the behavior is a bit
complicated (with mouse switching tab and mouse cancelling ongoing
action).

References #321