fix(core): re-used data if pipeline is trained
If the pipeline is trained, we call `predict`; if the pipeline is not
trained, we call `preProcessData`. In either case, we can directly use
the data; no need to process the data for the second time.

This fixes #337: If pipeline is trained, pipeline tab should use data
already flowing through the pipeline.