Update to latest ofxGrt.
This required some fixes, e.g. to link the ranges between channels
(which is no longer the default). Also, we need to keep pointers to
ofxGrtTimeseriesPlots now, instead of instances, because they can no
longer be copied because they contain a mutex. That means that the
feature plot shown in place of the input plot and the feature plot
shown at the bottom of the pipeline tab are the same plot, not two
copies, so we can only update one of them. We also need to set the
colors for all the ofxGrtTimeSeriesPlots.